// frontend/src/app/core/services/api/model/competitionGroupGender.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Gender classification for competition groups
 */
export const CompetitionGroupGender = {
    Male: 'MALE',
    Female: 'FEMALE',
    Mixed: 'MIXED'
} as const;
export type CompetitionGroupGender = typeof CompetitionGroupGender[keyof typeof CompetitionGroupGender];